Partial initial structure
Took a few steps towards setting up a structure for running multiple backends
that query PG via pgbouncer managed by Docker Compose.

Just for testing I am having each backend respond to `/ping` with the name of
the backend as well as the PG version (to verify we are querying PG).

PG is setup with two users:

- `postgres` - Superadmin
- `backend` - Read only access to the for the backends

PG is exposing it's port to the host to allow the data to be managed with the
superadmin acccount.

Within Docker, PG is only on the "database" network (along with pgbouncer) so
only pgbouncer can talk directly to the DB. Everyone else has to talk to
pg_bouncer.

Steps to take before we have at least a basic proof of concept:

* Add a load balancer (Traefik?) to rotate between the backends
* Put a skeleton Elixir code in the `/frontend` directory that will call the
  `/ping` endpoint on the load balancer
* Test out the load balancer by adding some "sleep" calls to ensure containers
  are taken in and out of rotation.

Once that is done I think we have a starter POC that we can iterate on.