Wrap up initial structure
This adds a temp frontend server in Elixir which queries a random backend
server and converts the JSON data to HTML. It also adds Traefik to both act
as an entrypoint by a browser request to be proxied off to the frontend but
also it will pick a random backend to serve Elixir's request.

Like the Crystal and Python backends, the Elixir frontend is just a placeholder
to verify the basic structure. I expect to to be replaced by something better
(probably a Phoenix-based site). Also I'm not packaging Elixir correctly in
docker. It needs to be more like Crystal where it builds in a temp container
and then just transfers the final binary or whatever to the runtime container.